refactor: 커스텀훅 개선
