+++ src/hooks/index.ts
@@ -5,3 +5,4 @@ export * from './useInfiniteScroll';
 export * from './useInfiniteScrollObserver';
 export * from './useIntersectionObserver';
 export * from './useLectures';
+export * from './usePagination';
